Fix: Allow reads consisting only of insertion + Fix CI

+++ src/genie/core/reference-source.cc
@@ -18,7 +18,7 @@ ReferenceSource::ReferenceSource(ReferenceManager* mgr) : refMgr(mgr) {}
 // ---------------------------------------------------------------------------------------------------------------------
 
 genie::core::meta::Reference ReferenceSource::getMeta() const {
-    return {"", 0, 0, 0, util::make_unique<meta::InternalRef>(0, 0), "Placeholder"};
+    return {"", 0, 0, 0, util::make_unique<meta::InternalRef>(uint16_t(0), uint16_t(0)), "Placeholder"};
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
